fs/{basic,verify}/tasks: add resource requirement hints
Add a hint to tasks that need more than 40GB disk per machine to
complete, as observed by running a job that turns out to fail because
the disk is full. These hints are interpreted by the teuthology OpenStack
backend.

Although these hints are likely to be more or less similar for all cloud
provisioning APIs, keep them OpenStack specific instead of attempting a
generalization out of nothing.

http://tracker.ceph.com/issues/12329 Refs: #12329

Signed-off-by: Loic Dachary <loic@dachary.org>